DRFtutorial/requirements.txt
asgiref==3.6.0
attrs==23.1.0
Django==4.1.9
django-cors-headers==3.14.0
django-filter==2.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
drf-yasg==1.21.5
Faker==18.7.0
inflection==0.5.1
jsonschema==4.17.3
psycopg2==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.2.4
uritemplate==4.1.1